+++ components/badge/demo/status.vue
@@ -10,6 +10,6 @@
     <Badge status="success" text="Success" />
     <Badge status="error" text="Error" />
     <Badge status="default" text="Default" />
-    <Badge status="warning" text="warning" />
+    <Badge status="warning" text="Warning" />
   </Space>
 </template>