+++ components/stat-number/demo/basic.vue
@@ -1,6 +1,12 @@
 <template>
   <Space vertical>
-    <Button @click="change" size="small">change</Button>
+    <Space wrap>
+      <Button @click="value = 12345" size="small">12,345</Button>
+      <Button @click="value = 54321" size="small">54,321</Button>
+      <Button @click="value -= 100" size="small">-100</Button>
+      <Button @click="value += 100" size="small">+100</Button>
+      <Button @click="change" size="small">change</Button>
+    </Space>
     <code>default:</code>
     <StatNumber v-model="value" />
     <code>type=`rollup`:</code>
@@ -15,7 +21,7 @@
 </template>
 <script setup lang="ts">
 import { ref } from "vue";
-const value = ref(Math.random() * 100000);
+const value = ref(12345);
 
 const change = () => {
   value.value = Math.random() * 10000000;